backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Marib Aldoais <marib.aldoais@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
logging_test = 'backend.logger:run'
validate_test = 'backend.validate_config:main'

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
logging = "^0.4.9.6"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
